requirements.txt
openai>=1.0.0
pyyaml>=6.0
requests>=2.25.1
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
arxiv>=2.1.0
pypdf>=5.0.0
PyMuPDF>=1.24.0
